Changes to structure, protocols and addition of bytebuffer/array
Restructured round a large majority of the code splitting class up. Such as the GameMessage Class alongside removing player from main and splitting it up into different classes.

+++ client/Project/core/src/main/kotlin/jackA/MaM/E4048541/BasePlayer.kt
@@ -7,7 +7,7 @@ import com.badlogic.gdx.graphics.glutils.ShapeRenderer
 // superclass - kotlins version of parent
 open class BasePlayer {
     // Our player
-    var ID : Int = 0 // Defaults to 0 gets set when player joins
+    var ID : Short = 0 // Defaults to 0 gets set when player joins
 
     var X : Float = 200f
     var Y : Float = 200f
